Fixed Crashes on Minimizing
Added a method to differentiate between minimizing and hiding an
application.  The minimization crashes occurred due to a specification
of the JVM.  If the last unthreaded main process is closed, the JVM will
close any JFrame objects that still exist, instead of minimizing.  The
current configuration appears to be poorly coded but it is simple a
feasible way around the JVM's constraints.