Recurring Tasks/Working Days Only/End Repeat creation working
Recurring tasks can now be created, and set to create on working days
only.  The end repeat date checkbox is now functioning properly.
Unfortunately, there is a known issue with the duplication of existing
methods.  If anyone can help me with this, I would appreciate that
because I am at a loss.  I will continue to work on it, but I wanted to
commit so as not to ruin working software.